cleanup mac build

+++ src/scripts/mkmacdmg.sh
@@ -10,21 +10,26 @@ if [ -z $APP_VERSION ]; then
     exit 1;
 fi
 
+if [ ! -f ../zcash/src/zcashd ]; then
+    echo "Could not find compiled zcashd in ../zcash/src/.";
+    exit 1;
+fi
+
 #Clean
-make distclean 2>&1 >/dev/null
+make distclean >/dev/null 2>&1
 
 # Build
 $QT_PATH/bin/qmake zec-qt-wallet.pro CONFIG+=release
 make -j4 >/dev/null
 
 #Qt deploy
-mkdir artifacts
-rm -f artifcats/zec-qt-wallet.dmg
-rm -f artifacts/rw*
+mkdir artifacts >/dev/null 2>&1
+rm -f artifcats/zec-qt-wallet.dmg >/dev/null 2>&1
+rm -f artifacts/rw* >/dev/null 2>&1
 cp ../zcash/src/zcashd zec-qt-wallet.app/Contents/MacOS/
 $QT_PATH/bin/macdeployqt zec-qt-wallet.app 
 
-create-dmg --volname "zec-qt-wallet-v0.3.0" --volicon "res/logo.icns" --window-pos 200 120 --icon "zec-qt-wallet.app" 200 190 --app-drop-link 600 185 --hide-extension "zec-qt-wallet.app"  --window-size 800 400 artifacts/zec-qt-wallet.dmg zec-qt-wallet.app
+create-dmg --volname "zec-qt-wallet-v0.3.0" --volicon "res/logo.icns" --window-pos 200 120 --icon "zec-qt-wallet.app" 200 190 --app-drop-link 600 185 --hide-extension "zec-qt-wallet.app"  --window-size 800 400 artifacts/zec-qt-wallet.dmg zec-qt-wallet.app >/dev/null
 
 mv artifacts/zec-qt-wallet.dmg artifacts/MacOS-zec-qt-wallet-v$APP_VERSION.dmg
